remove train tab
This tab has been broken since 6 months ago and no one even creates an issue for it ; this indicates no one uses it.
Also just tried original webui Train tab and it seems also broken even in Auto's repo.
so we remove it to make room for new tabs and new features.