Update style.css (#1698)
The previous change to the width of the input box, to prevent overlap between long checkpoint names and the dropdown arrow, also reduced the clickable region of the input box - showing up the arrow as the purely visual element that it is.
This PR restores width, tweaks position of the arrow, and makes the arrow background solid to avoid inference with text. The latter two changes also affect the vae/te and diffusion in low bits menus.